requirements.txt
appdirs==1.4.3
chardet==2.3.0
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.10.4
django-appconf==1.0.1
django-ckeditor==5.1.0
django-cors-headers==2.0.0
-e git://github.com/SalunMarvin/django-csvimport.git@94bb361ab056fb9d3327bc1017785b722f4e9066#egg=django_csvimport
django-filter==1.0.2
django-jet==1.0.3
django-selectable==0.9.0
djangorestframework==3.5.3
packaging==16.8
psycopg2==2.7.1
pyparsing==2.2.0
python-decouple==3.0
six==1.10.0
static3==0.7.0
